Remove raven-style key, better looping on batches
